% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebrel2-package.R
\name{generate_X0_CI_R}
\alias{generate_X0_CI_R}
\title{Generate naive starting point}
\usage{
generate_X0_CI_R(U, n_actions, dim_x, dim_y, base_prob = 0.85, seed)
}
\description{
Generate naive starting point
}
