% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebrel2-package.R
\docType{package}
\name{ebrel2-package}
\alias{ebrel2}
\alias{ebrel2-package}
\title{ebrel2: Ebrel2 - Spatial prioritisation for nature recovery and improvement}
\description{
R package for setting up and running a spatial prioritisation for habitat creation with multiple options and improvement of existing habitat. For technical details on the basic Ebrel model, see 10.1111/1365-2664.70144, and for further details on Ebrel2 see the vignette.
}
\author{
\strong{Maintainer}: David Baker \email{d.baker2@exeter.ac.uk} (\href{https://orcid.org/0000-0002-0466-8222}{ORCID})

Authors:
\itemize{
  \item David Baker \email{d.baker2@exeter.ac.uk} (\href{https://orcid.org/0000-0002-0466-8222}{ORCID})
}

}
\keyword{internal}
